requirements.txt
Django~=4.2.19
pdfkit~=1.0.0
qrcode~=8.0
djangorestframework~=3.15.2
jinja2~=3.1.5
pillow~=11.1.0